refactor: 使用 Copilot 重構並優化爬蟲和 API 資料抓取模組
- 重構 `web_scraper.py` 以提高代碼可讀性和可維護性。
- 添加註解以解釋每個方法和部分的用途。
- 更新 `api_data_fetcher.py` 以使用 `base_url` 作為初始化參數，增加靈活性。
- 添加錯誤處理和請求方法到 `api_data_fetcher.py`。
- 保留原始 `.env` 文件未做更改。
